Change to Static Libraryies and ZERO warnning.

+++ Desert/src/Platform/Windows/WindowsInput.cpp
@@ -28,7 +28,7 @@ namespace Desert {
 		double xpos, ypos;
 		glfwGetCursorPos(window, &xpos, &ypos);
 
-		return { xpos, ypos };
+		return { (float)xpos, (float)ypos };
 	}
 
 	float WindowsInput::GetMouseXImpl()
